+++ packages/middleware/src/composeMiddleware.js
@@ -1,0 +1,6 @@
+import { map, applyTo, compose } from 'ramda';
+
+const composeMiddleware = (...middleware) => reduxApi => next =>
+	compose(...map(applyTo(reduxApi), middleware))(next);
+
+export default composeMiddleware;